requirement.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.10
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-import-export==1.2.0
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
future==0.17.1
GoDaddyPy==2.2.7
httplib2==0.9.2
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
